packages/slack-bot-python/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudevents==1.9.0
contourpy==1.0.7
cycler==0.11.0
deprecation==2.1.0
Flask==2.2.3
fonttools==4.39.2
frozenlist==1.3.3
functions-framework==3.3.0
google-api-core==2.11.0
google-auth==2.16.3
google-cloud-core==2.3.2
google-cloud-pubsub==2.15.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.51.3
grpcio-status==1.51.3
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
multidict==6.0.4
numpy==1.24.2
openai==0.27.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
plotly==5.13.1
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.2
requests==2.28.2
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
slack-bolt==1.17.0
slack-sdk==3.20.2
tenacity==8.2.2
threadpoolctl==3.1.0
tqdm==4.65.0
urllib3==1.26.15
walk-up==0.1.0
watchdog==3.0.0
Werkzeug==2.2.3
yarl==1.8.2
